Kiran(AssignmentController)

+++ CaptoneProject.Services.AssignmentAPI/Controllers/AssignmentQuestionController.cs
@@ -1,9 +1,14 @@
 ﻿using AutoMapper;
+using Azure;
+using CaptoneProject.Services.AssignmentAPI.Data.Dto;
 using CaptoneProject.Services.AssignmentAPI.Data.Dto.Assignment;
 using CaptoneProject.Services.AssignmentAPI.Models;
 using CaptoneProject.Services.AssignmentAPI.Repository;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace CaptoneProject.Services.AssignmentAPI.Controllers
 {
@@ -13,16 +18,26 @@ namespace CaptoneProject.Services.AssignmentAPI.Controllers
     {
         private readonly IAssignmentQuestionRepository _assignmentQuestionRepository;
         private readonly IMapper _mapper;
+        private readonly ResponseDto _response;
         public AssignmentQuestionController(IAssignmentQuestionRepository assignmentQuestionRepository,IMapper mapper)
         {
             _assignmentQuestionRepository = assignmentQuestionRepository;
             _mapper = mapper;
+            _response= new ResponseDto();
         }
         [HttpPost]
-        public async Task<IActionResult> AddAssignment([FromBody] AssignmentQuestionDto assignmentDto,string trainerId)
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Roles = "TRAINER")]
+        public async Task<IActionResult> AddAssignment([FromBody] AssignmentQuestionDto assignmentDto)
         {
             try
             {
+                var trainerId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+                if (string.IsNullOrEmpty(trainerId))
+                {
+                    _response.IsSuccess = false;
+                    _response.Message = "Trainer Id Not found";
+                    return Unauthorized(_response);
+                }
                 var assignmentQuestion = _mapper.Map<AssignmentQuestion>(assignmentDto);
                 assignmentQuestion.TrainerId = trainerId;
                 await _assignmentQuestionRepository.AddAssignmentQuestion(assignmentQuestion);
@@ -51,16 +66,25 @@ namespace CaptoneProject.Services.AssignmentAPI.Controllers
             }
         }
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateAssignment(int id, [FromBody] AssignmentQuestionDto assignmentDto,[FromQuery] string trainerId)
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Roles = "TRAINER")]
+
+        public async Task<IActionResult> UpdateAssignment(int questionId, [FromBody] AssignmentQuestionDto assignmentDto)
         {
             try
             {
-                var existing = await _assignmentQuestionRepository.GetAssignmentQuestionById(id);
+                var trainerId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+                if (string.IsNullOrEmpty(trainerId))
+                {
+                    _response.IsSuccess = false;
+                    _response.Message = "Trainer Id Not found";
+                    return Unauthorized(_response);
+                }
+                var existing = await _assignmentQuestionRepository.GetAssignmentQuestionById(questionId);
                 if (existing == null) return NotFound("Assignment not found.");
 
                 var updated= _mapper.Map<Models.AssignmentQuestion>(assignmentDto);
                 updated.Id=existing.Id;
-                await _assignmentQuestionRepository.UpdateAssignmentQuestion(id,updated,trainerId);
+                await _assignmentQuestionRepository.UpdateAssignmentQuestion(questionId,updated,trainerId);
                 var response = _mapper.Map<AssignmentQuestionResponseDto>(updated);
                 return Ok(response);
             }
@@ -70,10 +94,18 @@ namespace CaptoneProject.Services.AssignmentAPI.Controllers
             }
         }
         [HttpDelete("{assignmentId}")]
-        public async Task<IActionResult> Delete(int assignmentId, [FromQuery] string trainerId)
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Roles = "TRAINER")]
+        public async Task<IActionResult> Delete(int assignmentId)
         {
             try
             {
+                var trainerId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+                if (string.IsNullOrEmpty(trainerId))
+                {
+                    _response.IsSuccess = false;
+                    _response.Message = "Trainer Id Not found";
+                    return Unauthorized(_response);
+                }
                 var deleted = await _assignmentQuestionRepository.DeleteAssignmentQuestion(trainerId,assignmentId);
                 if (!deleted) return NotFound("Assignment not found or access denied.");
                 return NoContent();
